third_party/WebKit/LayoutTests/http/tests/htmlimports/cors-same-origin.html

<!DOCTYPE html>
<html>
<head>
<link id="target" rel="import" href="http://localhost:8080/htmlimports/resources/cors-basic.cgi">
<script src="/js-test-resources/js-test-pre.js"></script>
</head>
<body>
<script>
shouldBe("target.import.querySelector('h1').innerHTML", "'Hello, CORS!'");
</script>
<script src="/js-test-resources/js-test-post.js"></script>
</body>
</html>
