third_party/WebKit/LayoutTests/http/tests/htmlimports/cors-same-origin-expected.txt

PASS target.import.querySelector('h1').innerHTML is 'Hello, CORS!'
PASS successfullyParsed is true

TEST COMPLETE
